# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ReliAPI",
    "description": "Reliability layer for API calls: retries, caching, dedup, circuit breakers, LLM proxy with cost control.",
    "version": "1.1.0",
    "contact": {
      "name": "KikuAI-Lab",
      "url": "https://github.com/kiku-jw/reliapi",
      "email": "dev@kikuai.dev"
    },
    "license": {
      "name": "AGPL-3.0-only",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    }
  },
  "paths": {
...
    "/v1/proxy/http": {
      "post": {
        "tags": [
          "Proxy"
        ],
        "summary": "Proxy HTTP request",
        "description": "Universal HTTP proxy endpoint for any HTTP API. Supports retries, circuit breaker, cache, and idempotency. Use this endpoint to add reliability layers to any HTTP API call.",
        "operationId": "proxy_http_v1_proxy_http_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HTTPProxyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/proxy/llm": {
      "post": {
        "tags": [
          "Proxy"
        ],
        "summary": "Proxy LLM request",
        "description": "LLM proxy endpoint with idempotency, budget caps, and caching. Make idempotent LLM API calls with predictable costs. Supports OpenAI, Anthropic, and Mistral providers. Set stream=true for Server-Sent Events (SSE) streaming.",
        "operationId": "proxy_llm_v1_proxy_llm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMProxyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/proxy/http": {
      "post": {
        "tags": [
          "Legacy",
          "Proxy"
        ],
        "summary": "Proxy HTTP request",
        "description": "Universal HTTP proxy endpoint for any HTTP API. Supports retries, circuit breaker, cache, and idempotency. Use this endpoint to add reliability layers to any HTTP API call.",
        "operationId": "proxy_http_proxy_http_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HTTPProxyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
    "/proxy/llm": {
      "post": {
        "tags": [
          "Legacy",
          "Proxy"
        ],
        "summary": "Proxy LLM request",
        "description": "LLM proxy endpoint with idempotency, budget caps, and caching. Make idempotent LLM API calls with predictable costs. Supports OpenAI, Anthropic, and Mistral providers. Set stream=true for Server-Sent Events (SSE) streaming.",
        "operationId": "proxy_llm_proxy_llm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LLMProxyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "deprecated": true
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPProxyRequest": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target",
            "description": "Target name from config.yaml (e.g., 'my_api')"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "HTTP method: GET, POST, PUT, DELETE, PATCH, HEAD, OPTIONS"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "API path (e.g., '/users/123' or '/api/v1/data')"
          },
          "headers": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headers",
            "description": "HTTP headers to include in request"
          },
          "query": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "Query parameters (e.g., {'page': 1, 'limit': 10})"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body",
            "description": "Request body as JSON string (for POST/PUT/PATCH)"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key",
            "description": "Idempotency key for request coalescing. Concurrent requests with same key execute once."
          },
          "cache": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache",
            "description": "Cache TTL in seconds (overrides config default). Only applies to GET/HEAD requests."
          }
        },
        "type": "object",
        "required": [
          "target",
          "method",
          "path"
        ],
        "title": "HTTPProxyRequest",
        "description": "Request schema for POST /proxy/http.\n\nUse this endpoint to proxy any HTTP API request with reliability layers:\n- Retries with exponential backoff\n- Circuit breaker per target\n- TTL cache for GET/HEAD requests\n- Idempotency with request coalescing"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMProxyRequest": {
        "properties": {
          "target": {
            "type": "string",
            "title": "Target",
            "description": "LLM target name from config.yaml (e.g., 'openai', 'anthropic')"
          },
          "messages": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "type": "array",
            "minItems": 1,
            "title": "Messages",
            "description": "Messages list with 'role' and 'content' (e.g., [{'role': 'user', 'content': 'Hello'}])"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Model name (e.g., 'gpt-4o-mini', 'claude-3-haiku'). Uses default from config if not specified."
          },
          "max_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens",
            "description": "Maximum tokens in response (limited by config max_tokens and budget caps)"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature",
            "description": "Temperature for sampling (0.0-2.0, limited by config)"
          },
          "top_p": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top P",
            "description": "Top-p sampling parameter (0.0-1.0)"
          },
          "stop": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop",
            "description": "Stop sequences (e.g., ['\\n', 'END'])"
          },
          "stream": {
            "type": "boolean",
            "title": "Stream",
            "description": "Streaming mode. If true, returns Server-Sent Events (SSE) stream. If false or omitted, returns standard JSON response.",
            "default": false
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key",
            "description": "Idempotency key for request coalescing. Use same key for duplicate requests to avoid duplicate LLM calls."
          },
          "cache": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache",
            "description": "Cache TTL in seconds (overrides config default). Cached responses return instantly without LLM call."
          }
        },
        "type": "object",
        "required": [
          "target",
          "messages"
        ],
        "title": "LLMProxyRequest",
        "description": "Request schema for POST /proxy/llm.\n\nMake idempotent LLM API calls with predictable costs.\nSupports OpenAI, Anthropic, and Mistral.\n\nFeatures:\n- Idempotency: duplicate requests return cached result\n- Budget caps: hard cap (reject) and soft cap (throttle)\n- Caching: TTL cache for LLM responses\n- Retries: automatic retries on failures"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}